make changes
"прокачав" таску 1, поправив глюки на локалхості з тасками 1 і 2 ( деколи перестає бачити файли js якщо в них назва починається з цифри)
