+++ package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "flatpickr": "^4.6.13",
+    "lodash.throttle": "^4.1.1",
     "modern-normalize": "^1.1.0",
     "notiflix": "^3.2.5"
   },
