+++ src/01-color-switcher.html
@@ -10,9 +10,18 @@
   <body>
     <p><a href="index.html">Go back</a></p>
     <div class="color-switcher">
-      <button type="button" class="color-switcher_btn" data-start>Start</button>
-      <button type="button" class="color-switcher_btn" data-stop>Stop</button>
+      <button type="button" class="color-switcher__btn" data-start>
+        Start
+      </button>
+      <button type="button" class="color-switcher__btn" data-stop>Stop</button>
+
+      <input id="change-bg-spead-control" type="range" min="0" max="100" />
+      <label for="change-bg-spead-control" class="color-switcher_label"
+        >speed of backgroung color change:
+        <span data-spead="50"></span> s</label
+      >
     </div>
-    <script src="js/01-color-switcher.js" type="module"></script>
+
+    <script src="js/color-switcher.js" type="module"></script>
   </body>
 </html>
