+++ src/02-timer.html
@@ -41,6 +41,6 @@
         </div>
       </div>
     </div>
-    <script src="js/02-timer.js" type="module"></script>
+    <script src="js/timer.js" type="module"></script>
   </body>
 </html>
