+++ src/css/common.css
@@ -11,7 +11,6 @@ body {
   color: #212121;
   line-height: 1.5;
   max-width: 1200px;
-  transition: background-color 250ms;
 }
 
 img {
